requirements.txt
python-telegram-bot==21.7
google-api-python-client==2.124.0
google-auth==2.29.0
google-auth-oauthlib==1.2.0
httpx==0.27.0
python-dotenv==1.0.1
nest_asyncio==1.5.8
